+++ lib/models/event_provider.dart
@@ -1,3 +1,4 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'event.dart';
@@ -22,6 +23,12 @@ final _events = [
   ),
 ];
 
-final eventsProvider = StreamProvider.autoDispose<List<Event>>((ref) async* {
-  yield _events;
+final eventsProvider = StreamProvider.autoDispose<List<Event>>((ref) {
+  return FirebaseFirestore.instance.collection('events').snapshots().map(
+        (snapshot) => snapshot.docs
+            .map(
+              (doc) => Event.fromJson(doc.data()),
+            )
+            .toList(),
+      );
 });