+++ lib/utils/bool.dart
@@ -1,0 +1,3 @@
+extension BoolExtension on bool {
+  T? nullIfFalse<T>(T value) => this ? value : null;
+}