+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## NEXT
 
 - Adds placeholder login and sign up screens.
+- Use Firestore to store events.
+- MVP to create events.
 
 ## 0.2.1 [2023-10-04]